Enhance testing with different etcd restart intervals

+++ networking_calico/tests/test_fv_etcdutils.py
@@ -31,6 +31,7 @@ eventlet.monkey_patch()
 from networking_calico.common import config as calico_config
 from networking_calico.compat import cfg
 from networking_calico import etcdutils
+from networking_calico import etcdv3
 
 _log = logging.getLogger(__name__)
 
@@ -38,6 +39,7 @@ _log = logging.getLogger(__name__)
 class TestFVEtcdutils(unittest.TestCase):
     def setUp(self):
         super(TestFVEtcdutils, self).setUp()
+        self.etcd_server_running = False
 
     def tearDown(self):
         self.stop_etcd_server()
@@ -48,16 +50,43 @@ class TestFVEtcdutils(unittest.TestCase):
                   " quay.io/coreos/etcd:v3.3.11 etcd" +
                   " --advertise-client-urls http://127.0.0.1:2379" +
                   " --listen-client-urls http://0.0.0.0:2379")
+        self.etcd_server_running = True
+
+    def wait_etcd_ready(self):
+        self.assertTrue(self.etcd_server_running)
+        ready = False
+        for ii in range(5):
+            try:
+                etcdv3.get_status()
+                ready = True
+                break
+            except Exception:
+                eventlet.sleep(1)
+        self.assertTrue(ready)
 
     def stop_etcd_server(self):
-        os.system("docker kill etcd")
+        if self.etcd_server_running:
+            os.system("docker kill etcd")
+        self.etcd_server_running = False
 
-    def test_restart_resilience(self):
+    def test_restart_resilience_2s(self):
+        self._test_restart_resilience(2)
+
+    def test_restart_resilience_5s(self):
+        self._test_restart_resilience(5)
+
+    def test_restart_resilience_15s(self):
+        self._test_restart_resilience(15)
+
+    def _test_restart_resilience(self, restart_interval_secs):
         # Start a real local etcd server.
         self.start_etcd_server()
 
         # Set up minimal config, so EtcdWatcher will use that etcd.
         calico_config.register_options(cfg.CONF)
+
+        # Ensure etcd server is ready.
+        self.wait_etcd_ready()
 
         # Create and start an EtcdWatcher.
         ew = etcdutils.EtcdWatcher('/calico/felix/v2/abc/host',
@@ -66,29 +95,35 @@ class TestFVEtcdutils(unittest.TestCase):
         ew.debug_reporter = lambda msg: debug_msgs.append(msg)
         eventlet.spawn(ew.start)
 
-        # Let it run for 10 seconds normally.
-        eventlet.sleep(10)
+        # Let it run for 5 seconds normally.  The EtcdWatcher writes a
+        # round-trip-check key every 3.3s (WATCH_TIMEOUT_SECS / 3), so
+        # 5s is enough for at least one of those writes.
+        eventlet.sleep(5)
 
         # Stop the etcd server.
         debug_msgs.append("Stopping etcd server")
         self.stop_etcd_server()
 
-        # Let it run for 10 seconds more.
-        eventlet.sleep(10)
+        # Wait for the specified restart interval.
+        eventlet.sleep(restart_interval_secs)
 
         # Restart the etcd server.
         debug_msgs.append("Restarting etcd server")
         self.start_etcd_server()
 
-        # Let it run for 10 seconds more.
-        eventlet.sleep(10)
+        # Ensure etcd server is ready.
+        self.wait_etcd_ready()
+
+        # Let it run for 5 seconds more.  As above, this should be
+        # enough for at least one round-trip-check key write.
+        eventlet.sleep(5)
 
         # Stop the EtcdWatcher.
         debug_msgs.append("Stopping EtcdWatcher")
         ew.stop()
 
         # Find the message for "Restarting etcd server" and count
-        # "Write round-trip key" messages before and after that.  Both
+        # "Wrote round-trip key" messages before and after that.  Both
         # counts should be non-zero if the EtcdWatcher is working
         # correctly before and after the etcd server restart.
         num_key_writes_before_restart = 0
@@ -97,7 +132,7 @@ class TestFVEtcdutils(unittest.TestCase):
         for msg in debug_msgs:
             if msg == "Restarting etcd server":
                 seen_restart_msg = True
-            if msg == "Write round-trip key":
+            if msg == "Wrote round-trip key":
                 if seen_restart_msg:
                     num_key_writes_after_restart += 1
                 else:
